create DoughnutChart component relates #30
