+++ client/src/components/Charts/index.js
@@ -1,3 +1,3 @@
 export { default as BarChart } from './BarChart';
 export { default as LineChart } from './LineChart';
-export { default as PieChart } from './PieChart';
+export { default as DoughnutChart } from './DoughnutChart';
